+++ src/main/scala/dxWDL/BValue.scala
@@ -178,5 +178,4 @@ object BValue {
             case other => throw new DeserializationException(s"BValue unexpected ${other}")
         }
     }
-
 }